Separated schema interfaces from makers and impls
