+++ src/schema/data-types/exports.ts
@@ -1,13 +1,2 @@
-export * from './allow-empty-string';
-export * from './any';
-export * from './array';
-export * from './boolean';
-export * from './custom';
-export * from './date';
-export * from './number';
-export * from './object';
-export * from './record';
-export * from './regex';
-export * from './restricted-number';
-export * from './string';
-export * from './tuple';
+export * from './makers/exports';
+export * from './types/exports';
