+++ src/schema/data-types/makers/any.ts
@@ -1,18 +1,12 @@
-import { getMeaningfulTypeof } from '../../type-utils/get-meaningful-typeof';
-import type { Schema } from '../../types/schema';
-import { InternalSchemaMakerImpl } from '../internal/internal-schema-maker-impl';
-import type { InternalValidationResult } from '../internal/types/internal-validation';
-import { cloner } from '../internal/utils/cloner';
-import { copyMetaFields } from '../internal/utils/copy-meta-fields';
-import { makeErrorResultForValidationMode } from '../internal/utils/make-error-result-for-validation-mode';
-import { makeClonedValueNoError } from '../internal/utils/make-no-error';
-import { supportVariableSerializationFormsForBooleanValues } from '../internal/utils/support-variable-serialization-forms-for-boolean-values';
-
-/** Requires a non-null, non-undefined value. */
-export interface AnySchema extends Schema {
-  schemaType: 'any';
-  clone: () => AnySchema;
-}
+import { getMeaningfulTypeof } from '../../../type-utils/get-meaningful-typeof';
+import { InternalSchemaMakerImpl } from '../../internal/internal-schema-maker-impl';
+import type { InternalValidationResult } from '../../internal/types/internal-validation';
+import { cloner } from '../../internal/utils/cloner';
+import { copyMetaFields } from '../../internal/utils/copy-meta-fields';
+import { makeErrorResultForValidationMode } from '../../internal/utils/make-error-result-for-validation-mode';
+import { makeClonedValueNoError } from '../../internal/utils/make-no-error';
+import { supportVariableSerializationFormsForBooleanValues } from '../../internal/utils/support-variable-serialization-forms-for-boolean-values';
+import type { AnySchema } from '../types/AnySchema';
 
 /** Requires a non-null, non-undefined value.  Use `allowNull` or `optional` if `null` or `undefined` values should also be allowed. */
 export const any = (): AnySchema => new AnySchemaImpl();
