+++ src/schema/data-types/makers/custom.ts
@@ -1,20 +1,20 @@
-import { getLogger } from '../../config/logging';
-import { getMeaningfulTypeof } from '../../type-utils/get-meaningful-typeof';
-import type { JsonValue } from '../../types/json-value';
-import type { Schema } from '../../types/schema';
-import type { SerDes } from '../../types/ser-des';
-import type { ValidationMode } from '../../types/validation-options';
-import { InternalSchemaMakerImpl } from '../internal/internal-schema-maker-impl';
-import type { InternalState } from '../internal/internal-schema-maker-impl/internal-state';
-import type { GenericContainer } from '../internal/types/generic-container';
-import type { InternalSchemaFunctions } from '../internal/types/internal-schema-functions';
-import type { InternalTransformationType, InternalValidationResult, InternalValidator } from '../internal/types/internal-validation';
-import type { LazyPath } from '../internal/types/lazy-path';
-import { cloner } from '../internal/utils/cloner';
-import { copyMetaFields } from '../internal/utils/copy-meta-fields';
-import { isErrorResult } from '../internal/utils/is-error-result';
-import { makeErrorResultForValidationMode } from '../internal/utils/make-error-result-for-validation-mode';
-import { makeClonedValueNoError, makeNoError } from '../internal/utils/make-no-error';
+import { getLogger } from '../../../config/logging';
+import { getMeaningfulTypeof } from '../../../type-utils/get-meaningful-typeof';
+import type { JsonValue } from '../../../types/json-value';
+import type { SerDes } from '../../../types/ser-des';
+import type { ValidationMode } from '../../../types/validation-options';
+import { InternalSchemaMakerImpl } from '../../internal/internal-schema-maker-impl';
+import type { InternalState } from '../../internal/internal-schema-maker-impl/internal-state';
+import type { GenericContainer } from '../../internal/types/generic-container';
+import type { InternalSchemaFunctions } from '../../internal/types/internal-schema-functions';
+import type { InternalTransformationType, InternalValidationResult, InternalValidator } from '../../internal/types/internal-validation';
+import type { LazyPath } from '../../internal/types/lazy-path';
+import { cloner } from '../../internal/utils/cloner';
+import { copyMetaFields } from '../../internal/utils/copy-meta-fields';
+import { isErrorResult } from '../../internal/utils/is-error-result';
+import { makeErrorResultForValidationMode } from '../../internal/utils/make-error-result-for-validation-mode';
+import { makeClonedValueNoError, makeNoError } from '../../internal/utils/make-no-error';
+import type { CustomSchema } from '../types/CustomSchema';
 
 export type CustomValidationResult = { error?: string } | { error?: undefined };
 
@@ -27,15 +27,6 @@ export interface CustomSchemaOptions<ValueT, SerializedT extends JsonValue> {
 
   /** Performs validation logic.  By default, only `isValueType` is checked, using the `serDes` field. */
   customValidation?: CustomValidation<ValueT>;
-}
-
-/** Used for adding custom schemas for complex types. */
-export interface CustomSchema<ValueT, SerializedT extends JsonValue> extends Schema<ValueT> {
-  schemaType: 'custom';
-  clone: () => CustomSchema<ValueT, SerializedT>;
-
-  serDes: SerDes<ValueT, SerializedT>;
-  typeName: string;
 }
 
 /**
