+++ src/schema/data-types/types/AnySchema.ts
@@ -1,0 +1,8 @@
+import type { Schema } from '../../../types/schema';
+
+/** Requires a non-null, non-undefined value. */
+
+export interface AnySchema extends Schema {
+  schemaType: 'any';
+  clone: () => AnySchema;
+}
