+++ src/schema/data-types/types/RegexSchema.ts
@@ -1,0 +1,10 @@
+import type { Schema } from '../../../types/schema';
+
+/** Requires a string matching the specified regular expression. */
+
+export interface RegexSchema extends Schema<string> {
+  schemaType: 'regex';
+  clone: () => RegexSchema;
+
+  regex: RegExp;
+}
