+++ src/schema/data-types/types/RestrictedNumberOptions.ts
@@ -1,0 +1,4 @@
+export interface RestrictedNumberOptions {
+  /** If one or more values are specified, the value must be divisible by one of the specified options */
+  divisibleBy?: number[];
+}
